# api/openapi.yaml
openapi: 3.0.3
info:
  title: Point of Service (POS) API
  description: REST API to be used to create tableside food ordering applications.
  contact:
    name: Course Discord server
    url: https://discord.gg/dPsnz5u9
    email: mark.wentowski@docsgeek.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: '1.0'
servers:
  - url: http://localhost:80/
    description: Local setup
  - url: '{gitpod-server-url}/'
    description: Gitpod setup
    variables:
      gitpod-server-url:
        default: 'https://8080-<workspace-id>.gitpod.io'
        description: 'Dynamic Gitpod URL. Replace <workspace-id> with your Gitpod workspace ID if needed.'
...
paths:
...
  /dishes:
    get:
      tags:
        - Dishes
      summary: Get a list of dishes
      description: |
        This operation retrieves a list of dishes based on the specified criteria. You can sort the results, include specific properties, select dishes that match certain criteria, and paginate through the results.
      operationId: getDishList
      parameters:
        - name: sort
          in: query
          description: |
            Specifies the field by which the results should be sorted. This parameter allows you to define the sorting order based on a particular field in the dataset. For example, if you want to sort the results by priority, you would use `priority` as the value for this parameter. If you include a `sort` value, you must include a `order` value.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: name
        - name: order
          in: query
          description: |
            Determines the order in which the results should be sorted. This parameter specifies whether the sorting should be in ascending or descending order. Valid values are `asc` for ascending order and `desc` for descending order. If you include an `order` value, you must include a `sort` value.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: asc
        - name: fields
          in: query
          description: |
            Use this parameter to specify which resource properties should be included in the response. Provide a list of schema property names, separated by the `|` delimiter. This allows for precise control over the data returned.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: name|ingredients|station
        - name: filter
          in: query
          description: |
            Selection filters to display resources matching the provided criteria. Use the format `fieldName.operator~value`.

            **Basic Equality (eq)**
            - For exact matches, use `fieldName.eq~value`.

            **Comparison Operators:**
            - **Less Than (lt)** Use `fieldName.lt~value`.
            - **Less Than or Equal To (lte)** Use `fieldName.lte~value`.
            - **Greater Than (gt):** Use `fieldName.gt~value`.
            - **Greater Than or Equal To (gte):** Use `fieldName.gte~value`.

            Separate filters using the `|` delimter. Dates should be in `YYYY-MM-DD` format.
          required: false
          style: form
          explode: true
          schema:
            type: string
            example: created_at.gt~2024-05-07|price.lte~19.99
        - name: limit
          in: query
          description: |
            Pagination filter for the maximum number of results to return in the response.
          required: false
          style: form
          explode: true
          schema:
            maximum: 30
            type: integer
          example: 10
        - name: offset
          in: query
          description: |
            Offest-based pagination filter for the number of results to skip before displaying resources.
          required: false
          style: form
          explode: true
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description:
            Successful operation. The response body contains the requested
            data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_3'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      x-swagger-router-controller: Dishes
    post:
      tags:
        - Dishes
      summary: Create a dish
      description: |
        This operation creates a new dish with the provided information. The `id` returned in the response can be used to associate the dish with an order.

        Each `ingredient_id` should refer to an ingredient currently in the system. For each ingredient, set the value of `is_essential` to `true` if the ingredient is required to prepare the dish. It's considered a best practice to ensure that each essential ingredient has a positive quantity. This helps determine if the dish should be selectable in the user interface.
      operationId: postDishes
      requestBody:
        $ref: '#/components/requestBodies/CreateDishRequest'
      responses:
        '201':
          description:
            Resource created successfully. The response body contains the
            newly created resource.
          headers:
            Location:
              description: Points to the resource endpoint for the newly created resource.
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description:
            The request could not be completed due to a conflict with the
            current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
        - AccessToken: []
      x-swagger-router-controller: Dishes
  /dishes/{id}:
    get:
      tags:
        - Dishes
      summary: Get a dish's details
      description: |
        Retrieve detailed information about a specific dish identified by its unique dish ID.
      operationId: getDish
      parameters:
        - name: id
          in: path
          description: The unique identifier for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 66294b2a4475a41f3e709bd1
        - name: fields
          in: query
          description: |
            Projection filters for including only certain resource properties in the response. The value for the parameter should be a comma-separated list of properties from the schema.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: name|price|station|ingredients
      responses:
        '200':
          description:
            Successful operation. The response body contains the requested
            data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      x-swagger-router-controller: Dishes
    put:
      tags:
        - Dishes
      summary: Update a dish
      description: |
        This operation updates an existing dish with the provided data.
      operationId: putDish
      parameters:
        - name: id
          in: path
          description: The unique identifier for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 66294b2a4475a41f3e709bce
      requestBody:
        $ref: '#/components/requestBodies/UpdateDishRequest'
      responses:
        '200':
          description:
            Resource updated successfully. The response body contains the
            updated resource.
          headers:
            Location:
              description: Points to the resource endpoint for the newly created resource.
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        '204':
          description:
            "No content. The operation was successful, and there is no\
            \ content to return."
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description:
            The request could not be completed due to a conflict with the
            current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
        - AccessToken: []
      x-swagger-router-controller: Dishes
    delete:
      tags:
        - Dishes
      summary: Delete a dish
      description:
        This operation deletes an existing dish identified by its unique
        dish ID.
      operationId: deleteDish
      parameters:
        - name: id
          in: path
          description: The unique identifier for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 66294b2a4475a41f3e709bd0
      responses:
        '204':
          description:
            "No content. The operation was successful, and there is no\
            \ content to return."
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
        - AccessToken: []
      x-swagger-router-controller: Dishes
  /dishes/{id}/ingredients:
    get:
      tags:
        - Dishes
      summary: Get a dish's ingredients
      description:
        "This operation retrieves a list of ingredients belonging to a\
        \ specific dish. Importantly, the `in_stock_qty` for a dish indicates how\
        \ much of the ingredient is available."
      operationId: getDishIngredients
      parameters:
        - name: id
          in: path
          description: The unique identifier for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 66294b2a4475a41f3e709bd2
        - name: sort
          in: query
          description: |
            Specifies the field by which the results should be sorted. This parameter allows you to define the sorting order based on a particular field in the dataset. If you include a `sort` value, you must include a `order` value.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: name
        - name: order
          in: query
          description: |
            Determines the order in which the results should be sorted. This parameter specifies whether the sorting should be in ascending or descending order. Valid values are `asc` for ascending order and `desc` for descending order. If you include an `order` value, you must include a `sort` value.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: asc
        - name: fields
          in: query
          description: |
            Use this parameter to specify which resource properties should be included in the response. Provide a list of schema property names, separated by the `|` delimiter. This allows for precise control over the data returned.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: name|in_stock_qty
        - name: filter
          in: query
          description: |
            Selection filters to display resources matching the provided criteria. Use the format `fieldName.operator~value`.

            **Basic Equality (eq)**
            - For exact matches, use `fieldName.eq~value`.

            **Comparison Operators:**
            - **Less Than (lt)** Use `fieldName.lt~value`.
            - **Less Than or Equal To (lte)** Use `fieldName.lte~value`.
            - **Greater Than (gt):** Use `fieldName.gt~value`.
            - **Greater Than or Equal To (gte):** Use `fieldName.gte~value`.

            Separate filters using the `|` delimter. Dates should be in `YYYY-MM-DD` format.
          required: false
          style: form
          explode: true
          schema:
            type: string
            example: created_at.gt~2024-05-07|in_stock_qty.gt~3
        - name: limit
          in: query
          description: |
            Pagination filter for the maximum number of results to return in the response.
          required: false
          style: form
          explode: true
          schema:
            maximum: 30
            type: integer
          example: 10
        - name: offset
          in: query
          description: |
            Offest-based pagination filter for the number of results to skip before displaying resources.
          required: false
          style: form
          explode: true
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description:
            Successful operation. The response body contains the requested
            data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_4'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      x-swagger-router-controller: Dishes
...
components:
  schemas:
...
    Category:
      type: string
      description:
        A category that organizes meal courses for display in food ordering
        apps.
      example: Main Course
      enum:
        - Appetizer
        - Main Course
        - Dessert
        - Drinks
    CreatedAt:
      type: string
      description: Timestamp when the resource was created in ISO 8601 format.
      format: date-time
      readOnly: true
      example: 2024-04-22T10:00:00Z
    Description:
      type: string
      example: A description of the resource.
    Dish:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        created_at:
          $ref: '#/components/schemas/CreatedAt'
        name:
          type: string
          description: A human-friendly name for the dish.
          example: Risotto alla Milanese
        description:
          $ref: '#/components/schemas/Description'
        category:
          $ref: '#/components/schemas/Category'
        image_name:
          $ref: '#/components/schemas/ImageName'
        ingredients:
          $ref: '#/components/schemas/Ingredients'
        preparation_time:
          $ref: '#/components/schemas/Price'
        price:
          $ref: '#/components/schemas/Price'
        station:
          $ref: '#/components/schemas/Station'
      example:
        image_name: burger
        price: null
        name: Risotto alla Milanese
        station: hot
        created_at: 2024-04-22T10:00:00Z
        description: A description of the resource.
        ingredients:
          - - ingredient_id: 66294b2a4475a41f3e709bc5
              is_essential: true
            - ingredient_id: 66294b2a4475a41f3e709bc6
              is_essential: true
            - ingredient_id: 66294b2a4475a41f3e709bc7
              is_essential: false
          - - ingredient_id: 66294b2a4475a41f3e709bc5
              is_essential: true
            - ingredient_id: 66294b2a4475a41f3e709bc6
              is_essential: true
            - ingredient_id: 66294b2a4475a41f3e709bc7
              is_essential: false
        preparation_time: 999
        id: 674d0bf5c28b69001f8e03a1
        category: Main Course
...
    Id:
      type: string
      description: The unique identifier for the resource.
      readOnly: true
      example: 674d0bf5c28b69001f8e03a1
    ImageName:
      type: string
      description: |
        Use this string to generate the asset URL for displaying the dish image in user interfaces.
      example: burger
    Ingredients:
      type: array
      description: List of ingredients used during the preparation of the dish.
      items:
        $ref: '#/components/schemas/Ingredients_inner'
    Ingredient:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        created_at:
          $ref: '#/components/schemas/CreatedAt'
        updated_at:
          $ref: '#/components/schemas/UpdatedAt'
        name:
          type: string
          description: A human-friendly name for the ingredient.
          example: Carrot
        in_stock_qty:
          $ref: '#/components/schemas/InStockQty'
        price:
          $ref: '#/components/schemas/Price'
      example:
        updated_at: 2024-04-22T10:00:00Z
        price: 999
        name: Carrot
        created_at: 2024-04-22T10:00:00Z
        id: 674d0bf5c28b69001f8e03a1
        in_stock_qty: 3
    InStockQty:
      type: integer
      description:
        The quantity of the ingredient in the restaurant's inventory. The
        quantity of the ingredient should be managed by the API consumer.
      example: 3
...
    PreparationTime:
      type: integer
      description: 'Estimated time required to prepare the dish, in minutes.'
      example: 45
    Price:
      type: number
      description:
        The cost of the dish in cents. For example `999` equals `9.99`
        and `12999` equals `129.99`.
      format: float
      example: 999
...
    Problem:
      required:
        - detail
        - title
      type: object
      properties:
        title:
          type: string
          description: Human-readable error title.
          example: Human-readable error title.
        detail:
          type: string
          description: Human-readable error details.
          example: Human-readable error details.
...
    Station:
      type: string
      description: |
        The restaurant kitchen has 3 stations where food is prepared:
          - cold: This area is responsible for preparing cold appetizers, salads, and desserts that don't require cooking.
          - hot: Also known as the "hot line" or "hot side," this area handles dishes that require cooking over heat, such as sautéing, frying, grilling, and baking.
          - beverages: In addition to coffee bars, there may be stations dedicated to preparing other beverages such as cocktails, smoothies, or fresh juices.
      example: hot
      enum:
        - cold
        - hot
        - beverages
...
    UpdatedAt:
      type: string
      description: Timestamp when the order was created in ISO 8601 format.
      format: date-time
      readOnly: true
      example: 2024-04-22T10:00:00Z
...
    inline_response_200_3:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Dish'
        total_results:
          type: integer
          description: The number of resources returned in the response.
          example: 20
      example:
        results:
          - image_name: burger
            price: null
            name: Risotto alla Milanese
            station: hot
            created_at: 2024-04-22T10:00:00Z
            description: A description of the resource.
            ingredients:
              - - ingredient_id: 66294b2a4475a41f3e709bc5
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc6
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc7
                  is_essential: false
              - - ingredient_id: 66294b2a4475a41f3e709bc5
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc6
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc7
                  is_essential: false
            preparation_time: 999
            id: 674d0bf5c28b69001f8e03a1
            category: Main Course
          - image_name: burger
            price: null
            name: Risotto alla Milanese
            station: hot
            created_at: 2024-04-22T10:00:00Z
            description: A description of the resource.
            ingredients:
              - - ingredient_id: 66294b2a4475a41f3e709bc5
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc6
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc7
                  is_essential: false
              - - ingredient_id: 66294b2a4475a41f3e709bc5
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc6
                  is_essential: true
                - ingredient_id: 66294b2a4475a41f3e709bc7
                  is_essential: false
            preparation_time: 999
            id: 674d0bf5c28b69001f8e03a1
            category: Main Course
        total_results: 20
    inline_response_200_4:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        total_results:
          type: integer
          example: 1
      example:
        results:
          - updated_at: 2024-04-22T10:00:00Z
            price: 999
            name: Carrot
            created_at: 2024-04-22T10:00:00Z
            id: 674d0bf5c28b69001f8e03a1
            in_stock_qty: 3
          - updated_at: 2024-04-22T10:00:00Z
            price: 999
            name: Carrot
            created_at: 2024-04-22T10:00:00Z
            id: 674d0bf5c28b69001f8e03a1
            in_stock_qty: 3
        total_results: 1
...
    Ingredients_inner:
      type: object
      properties:
        ingredient_id:
          type: string
          description: The unique identifier for the resource.
          example: 66294b2a4475a41f3e709bc4
        is_essential:
          type: boolean
          description: |
            Whether the ingredient is essential to prepare the dish.
          example: true
      example:
        - ingredient_id: 66294b2a4475a41f3e709bc5
          is_essential: true
        - ingredient_id: 66294b2a4475a41f3e709bc6
          is_essential: true
        - ingredient_id: 66294b2a4475a41f3e709bc7
          is_essential: false
  requestBodies:
    CreateDishRequest:
      description: Request body containing information used to create the resource.
      content:
        application/json:
          schema:
            required:
              - category
              - description
              - image_name
              - ingredients
              - name
              - preparation_time
              - price
              - station
            type: object
            properties:
              name:
                type: string
                description: A human-friendly name for the dish.
                example: Risotto alla Milanese
              description:
                $ref: '#/components/schemas/Description'
              category:
                $ref: '#/components/schemas/Category'
              image_name:
                $ref: '#/components/schemas/ImageName'
              ingredients:
                $ref: '#/components/schemas/Ingredients'
              preparation_time:
                $ref: '#/components/schemas/PreparationTime'
              price:
                $ref: '#/components/schemas/Price'
              station:
                $ref: '#/components/schemas/Station'
      required: true
    UpdateDishRequest:
      description: Request body containing information used to update the resource.
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
                description: A human-friendly name for the dish.
                example: Risotto alla Milanese
              description:
                $ref: '#/components/schemas/Description'
              category:
                $ref: '#/components/schemas/Category'
              image_name:
                $ref: '#/components/schemas/ImageName'
              ingredients:
                $ref: '#/components/schemas/Ingredients'
              preparation_time:
                $ref: '#/components/schemas/PreparationTime'
              price:
                $ref: '#/components/schemas/Price'
              station:
                $ref: '#/components/schemas/Station'
      required: true
...
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
    BasicAuth:
      type: http
      description: |
        Authentication using client credentials. Use the `client_id` as the username and `client_secret` as the password.

        - **client_id**: Public identifier for the client application. Example: `admin_web_app`.
        - **client_secret**: Confidential string for client authentication. Example: `a5d7f23e-8b64-4b4c-9b11-21c5cfdf25f1`.
      scheme: basic